requirements.txt
psycopg2-binary
pandas
selenium
webdriver-manager
matplotlib
seaborn
numpy
scipy
scikit-learn
tensorflow